03-spring-data-jpa/_01-persistence-context-in-spring/src/main/java/com/sh/app/menu/repository/MenuRepository.java
package com.sh.app.menu.repository;

import com.sh.app.menu.entity.Menu;
import jakarta.persistence.EntityManager;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;

@Repository
@RequiredArgsConstructor
public class MenuRepository {
    private final EntityManager entityManager;

    public Menu findByMenuCode(Long menuCode) {
        return entityManager.find(Menu.class, menuCode);
    }

    public void updateMenuPriceBy10Won(Long menuCode) {
        Menu menu = entityManager.find(Menu.class, menuCode);
        menu.setMenuPrice(menu.getMenuPrice() + 10);
    }
}
